chore: refactorise `TaskDurationBar` and move part of it in newly created `ProfilingDurationBar`
